refactor: rename git_operations -> git, GitContext -> Git
Module path `git_machete.git_operations` is renamed to `git_machete.git`,
the central `GitContext` class to `Git`, and the test file
`tests/test_git_operations.py` to `tests/test_git.py`. Pure rename - no
behavioural change.

The previous names predated the package layout cleanup; the new ones are
shorter, less verbose, and consistent with how callers already say
`from git_machete.git import ...` after the recent refactor of `utils`.
The mocking whitelist and one test patch path are updated to match.
