+++ ci/checks/enforce-mocking-only-whitelisted-methods.sh
@@ -11,7 +11,7 @@ builtins.input
 git_machete.client.go_interactive.GoInteractiveMacheteClient._get_stdin_fd
 git_machete.client.go_interactive.GoInteractiveMacheteClient._read_stdin
 git_machete.code_hosting.OrganizationAndRepository.from_url
-git_machete.git_operations.GitContext.fetch_remote
+git_machete.git.Git.fetch_remote
 git_machete.github.GitHubClient.MAX_PULLS_PER_PAGE_COUNT
 git_machete.github.GitHubToken.for_domain
 git_machete.gitlab.GitLabClient.MAX_PULLS_PER_PAGE_COUNT
