+++ git_machete/client/advance.py
@@ -1,13 +1,12 @@
 from git_machete.client.base import MacheteClient
 from git_machete.config import SquashMergeDetection
-from git_machete.git_operations import (GitContext, LocalBranchShortName,
-                                        SyncToRemoteStatus)
+from git_machete.git import Git, LocalBranchShortName, SyncToRemoteStatus
 from git_machete.utils.exceptions import MacheteException
 from git_machete.utils.markup import pretty_choices, warn
 
 
 class AdvanceMacheteClient(MacheteClient):
-    def __init__(self, git: GitContext) -> None:
+    def __init__(self, git: Git) -> None:
         super().__init__(git)
 
     def advance(self, *, opt_yes: bool) -> None:
