+++ git_machete/client/diff.py
@@ -2,7 +2,7 @@ import os
 from typing import List, Optional
 
 from git_machete.client.base import MacheteClient
-from git_machete.git_operations import LocalBranchShortName
+from git_machete.git import LocalBranchShortName
 
 
 class DiffMacheteClient(MacheteClient):
