+++ git_machete/client/discover.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Tuple
 from git_machete.client.status import StatusMacheteClient
 from git_machete.config import SquashMergeDetection
 from git_machete.constants import DISCOVER_DEFAULT_FRESH_BRANCH_COUNT
-from git_machete.git_operations import LocalBranchShortName
+from git_machete.git import LocalBranchShortName
 from git_machete.utils.collections import excluding, tupled
 from git_machete.utils.debug_log import debug
 from git_machete.utils.exceptions import MacheteException
