+++ git_machete/client/fork_point.py
@@ -1,5 +1,5 @@
 from git_machete.client.base import MacheteClient
-from git_machete.git_operations import AnyRevision, LocalBranchShortName
+from git_machete.git import AnyRevision, LocalBranchShortName
 from git_machete.utils.exceptions import MacheteException
 from git_machete.utils.markup import print_fmt
