+++ git_machete/client/slide_out.py
@@ -2,7 +2,7 @@ from typing import List, Optional
 
 from git_machete.client.base import MacheteClient
 from git_machete.config import SquashMergeDetection
-from git_machete.git_operations import AnyRevision, LocalBranchShortName
+from git_machete.git import AnyRevision, LocalBranchShortName
 from git_machete.utils.exceptions import MacheteException
 from git_machete.utils.markup import green_ok, print_fmt
